requirements.txt
pyTelegramBotAPI
Flask
pytesseract
Pillow
gunicorn
